Fix similarItems circular ref re: gracefulVariants
